+++ tests/test_config.py
@@ -4,11 +4,17 @@ from unittest import mock
 
 import pytest
 
-from openeo_aggregator.config import AggregatorConfig, STREAM_CHUNK_SIZE_DEFAULT, PRODUCTION_CONFIG
-from openeo_aggregator.config import OPENEO_AGGREGATOR_CONFIG, ENVIRONMENT_INDICATOR
-from openeo_aggregator.config import DEVELOPMENT_CONFIG, get_config
+from openeo_aggregator.config import AggregatorConfig, STREAM_CHUNK_SIZE_DEFAULT, \
+    OPENEO_AGGREGATOR_CONFIG, ENVIRONMENT_INDICATOR, get_config
 
-CONFIG_JSON_EXAMPLE = '{"aggregator_backends":{"b1":"https://b1.test"},"streaming_chunk_size":123}'
+CONFIG_PY_EXAMPLE = """
+from openeo_aggregator.config import AggregatorConfig
+config = AggregatorConfig(
+    config_source=__file__,
+    aggregator_backends={"b1": "https://b1.test"},
+    streaming_chunk_size=123
+)
+"""
 
 
 def test_config_defaults():
@@ -26,18 +32,12 @@ def test_config_aggregator_backends():
     assert config.aggregator_backends == {"b1": "https://b1.test"}
 
 
-def test_config_from_json():
-    data = '{"aggregator_backends": {"b1": "https://b1.test"}, "streaming_chunk_size": 123}'
-    config = AggregatorConfig.from_json(data)
-    assert config.aggregator_backends == {"b1": "https://b1.test"}
-    assert config.streaming_chunk_size == 123
-
-
-def test_config_from_json_file(tmp_path):
-    config_path = tmp_path / "config.json"
-    with config_path.open("w") as f:
-        f.write('{"aggregator_backends": {"b1": "https://b1.test"}, "streaming_chunk_size": 123}')
-    config = AggregatorConfig.from_json_file(str(config_path))
+def test_config_from_py_file(tmp_path):
+    path = tmp_path / "aggregator-conf.py"
+    with path.open(mode="w") as f:
+        f.write(CONFIG_PY_EXAMPLE)
+    config = AggregatorConfig.from_py_file(path)
+    assert config.config_source == str(path)
     assert config.aggregator_backends == {"b1": "https://b1.test"}
     assert config.streaming_chunk_size == 123
 
@@ -46,38 +46,28 @@ def test_get_config_none_no_env():
     assert OPENEO_AGGREGATOR_CONFIG not in os.environ
     assert ENVIRONMENT_INDICATOR not in os.environ
     config = get_config(None)
-    assert config is DEVELOPMENT_CONFIG
-
-
-def test_get_config_json_str():
-    config = get_config(CONFIG_JSON_EXAMPLE)
-    assert config.aggregator_backends == {"b1": "https://b1.test"}
-    assert config.streaming_chunk_size == 123
+    assert config.config_source.endswith("/conf/aggregator.dev.py")
 
 
 @pytest.mark.parametrize("convertor", [str, Path])
-def test_get_config_json_path(tmp_path, convertor):
-    config_path = tmp_path / "config.json"
+def test_get_config_py_file_path(tmp_path, convertor):
+    config_path = tmp_path / "aggregator-conf.py"
     with open(config_path, "w") as f:
-        f.write(CONFIG_JSON_EXAMPLE)
+        f.write(CONFIG_PY_EXAMPLE)
     config = get_config(convertor(config_path))
+    assert config.config_source == str(config_path)
     assert config.aggregator_backends == {"b1": "https://b1.test"}
     assert config.streaming_chunk_size == 123
 
 
-def test_get_config_none_env_json_str():
-    with mock.patch.dict(os.environ, {OPENEO_AGGREGATOR_CONFIG: CONFIG_JSON_EXAMPLE}):
-        config = get_config(None)
-    assert config.aggregator_backends == {"b1": "https://b1.test"}
-    assert config.streaming_chunk_size == 123
+def test_get_config_env_py_file(tmp_path):
+    path = tmp_path / "aggregator-conf.py"
+    with path.open(mode="w") as f:
+        f.write(CONFIG_PY_EXAMPLE)
 
-
-def test_get_config_none_env_json_path(tmp_path):
-    config_path = tmp_path / "config.json"
-    with open(config_path, "w") as f:
-        f.write(CONFIG_JSON_EXAMPLE)
-    with mock.patch.dict(os.environ, {OPENEO_AGGREGATOR_CONFIG: str(config_path)}):
+    with mock.patch.dict(os.environ, {OPENEO_AGGREGATOR_CONFIG: str(path)}):
         config = get_config(None)
+    assert config.config_source == str(path)
     assert config.aggregator_backends == {"b1": "https://b1.test"}
     assert config.streaming_chunk_size == 123
 
@@ -86,19 +76,11 @@ def test_get_config_none_env_json_path(tmp_path):
 def test_get_config_none_env_dev(env):
     with mock.patch.dict(os.environ, {ENVIRONMENT_INDICATOR: env}):
         config = get_config(None)
-    assert config is DEVELOPMENT_CONFIG
+    assert config.config_source.endswith("/conf/aggregator.dev.py")
 
 
 @pytest.mark.parametrize("env", ["prod", "PROD"])
 def test_get_config_none_env_prod(env):
     with mock.patch.dict(os.environ, {ENVIRONMENT_INDICATOR: env}):
         config = get_config(None)
-    assert config is PRODUCTION_CONFIG
-
-
-def test_get_config_json_url_encoded():
-    # from: `import urllib.parse; urllib.parse.quote('{"aggregator_backends":{"b1":"https://b1.test"},"streaming_chunk_size":123}')`
-    data = '%7B%22aggregator_backends%22%3A%7B%22b1%22%3A%22https%3A//b1.test%22%7D%2C%22streaming_chunk_size%22%3A123%7D'
-    config = get_config(data)
-    assert config.aggregator_backends == {"b1": "https://b1.test"}
-    assert config.streaming_chunk_size == 123
+    assert config.config_source.endswith("/conf/aggregator.prod.py")